Khan Academy Playlist Project
I had to begin by writing a query that brought up all the songs by Queen. 
The next step was then to write a query that would select all the artist from the "Pop" genre.
Lastly it required a query that entailed a subquery where it would select all the songs that had artist in the "Pop" genre.